Relation in set
Suppose that A = {1, 2, 3} and B = {1, 2}. Let R be the relation from A to B containing (a, b) if
a∈A, b∈B, and a > b. Write a program to find the relation R and also represent this relation in
matrix form if a1 = 1, a2 = 2, and a3 = 3, and b1 = 1 and b2 = 2.